multiplier-encoder-master/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
Deprecated==1.2.14
dimod==0.12.6
diskcache==5.6.1
dwave-cloud-client==0.10.6
dwave-greedy==0.3.0
dwave-hybrid==0.6.10
dwave-inspector==0.4.2
dwave-neal==0.6.0
dwave-networkx==0.8.14
dwave-ocean-sdk==6.4.1
dwave-preprocessing==0.5.4
dwave-samplers==1.0.0
dwave-system==1.19.0
dwave-tabu==0.5.0
dwavebinarycsp==0.2.0
fasteners==0.18
Flask==2.3.2
fonttools==4.42.1
homebase==1.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.2
minorminer==0.2.11
networkx==2.8.8
numpy==1.25.2
packaging==23.1
parameterized==0.9.0
penaltymodel==1.0.2
Pillow==10.0.0
plucky==0.4.3
pydantic==1.10.12
pyparsing==2.4.7
pyqubo==1.4.0
PySMT==0.9.5
PySocks==1.7.1
python-dateutil==2.8.2
rectangle-packer==2.0.1
requests==2.31.0
scipy==1.11.1
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
wrapt==1.15.0
z3-solver==4.12.2.0
